corrected naming conventions between AdminElection and AdminNewElection
